security reset password

+++ src/main/java/uz/pdp/hotelsystem/repository/BookingRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uz.pdp.hotelsystem.entity.BookingRoom;
 
 
-public interface BookingRepository extends JpaRepository<BookingRoom,Integer> {
+public interface BookingRepository extends JpaRepository<BookingRoom,Long> {
 }
